requirements.txt
pdfplumber==0.11.6
requests==2.32.3
python-dotenv==1.1.0
flask==3.1.0
flask-cors==5.0.1
google-api-python-client==2.166.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
